CI: remove /etc/yum.conf symlink to allow clean yum reinstall
Due to some change in the mock the /etc/yum.conf file became symlink
to the /etc/yum/yum.conf. We then ran into the race condition because
the re-install of the yum wasn't replacing the /etc/yum.conf, and once
we removed the /etc/yum/yum.conf that the symlink was pointing to, the
yum stopped functioning properly because it was missing the default
configuration.

Change-Id: I2323d56025b3715554455e6f9ee46e4db2ff64ee
Signed-off-by: Lev Veyde <lveyde@redhat.com>
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>
(cherry picked from commit 4086fc9ef2f951ce0ba7b0abaefa615081ca6751)