build: Make resulting RPMs architecture-specific
ovirt-host[-depencies] where built as noarch RPMs. This wasn't
correct though, because the dependencies are in fact architecture
specific.

For instance 'dmidecode' is not available on all architecture, and
the specfile already had a condition dependency in place.
It can easily be shown that the generated RPM doesn't contain
a dependency on 'dmidecode' even on x86.

This specfile change removes the BuildArch: noarch statement
making sure architecture specific RPMs are build.
Additionaly the Obsoletes statement makes sure that the old
noarch RPMs are removed on a dnf/yum update, and a Conflicts
statement prevents that both noarch and arch-specific RPMs are
installed in parallel.

Change-Id: I872635370ef8ec637f8056a6315508138a28f2fc
Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>
(cherry picked from commit f8f4548bec2ceaa23c3b9a22fa368531e163c829)