copr: enable automatic build on merge
An automation has been set in place for ovirt-host package
on https://copr.fedorainfracloud.org/coprs/ovirt/ovirt-master-snapshot/

The .copr/Makefile script will be called to generate a src.rpm which
will then be used by copr to perform the build.

Result will be made available as a badge in the README.md file.

Change-Id: I754afd6637870dbb5d929ce282d7a42009dde487
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>